requirements.txt
livekit>=1.0
livekit-agents[openai,deepgram,cartesia,silero,turn_detector,assemblyai]~=1.2
livekit-plugins-noise-cancellation~=0.2
python-dotenv~=1.0
